+++ src/screens/PlaybackProblemScreen.tsx
@@ -1,7 +1,8 @@
 import { Text, useTheme } from "react-native-paper";
 import { SafeAreaView } from "react-native-safe-area-context";
 import { Dimensions, Pressable, StyleSheet, View } from "react-native";
-import Icon from "react-native-vector-icons/Ionicons";
+import IonIcon from "react-native-vector-icons/Ionicons";
+import AntIcon from 'react-native-vector-icons/AntDesign';
 import { useNavigation } from "@react-navigation/native";
 import { NativeStackNavigationProp } from "@react-navigation/native-stack";
 import { RootStackParams } from "navigators/RootNavigator";
@@ -21,21 +22,30 @@ const PlaybackProblemScreen = () => {
     const { nowPlaying, playBook, pauseBook } = useContext(PlaybackContext)
 
     const attemptReset = () => {
-        if (nowPlaying) {
-            playBook(nowPlaying)
+        if (nowPlaying && nowPlaying.tracks) {
+            console.log(`Resetting playback for book: ${nowPlaying.name}`)
+            playBook(nowPlaying, nowPlaying.tracks, { reset: true })
         }
         navigation.navigate('Core');
     }
 
     return (
         <SafeAreaView style={{ ...globalStyles.BGColor, ...styles.ScreenContainer }}>
+            <Pressable
+                style={styles.BackButton}
+                onPress={() => {
+                    navigation.goBack()
+                }}
+            >
+                <AntIcon name="left" size={24} color={theme.colors.primary} />
+            </Pressable>
             <View style={styles.ContentContainer}>
                 <View>
                     <Text variant="displayMedium">Playback Error</Text>
                     <Text style={{ marginTop: 10 }} variant="bodyLarge">This can happen when your network connection is interrupted.</Text>
                 </View>
                 <Pressable style={{ alignItems: 'center', marginTop: 60 }} onPress={attemptReset}>
-                    <Icon name="reload" size={64} color={primaryColor} />
+                    <IonIcon name="reload" size={64} color={primaryColor} />
                     <Text variant='bodyMedium'>Reset Playback</Text>
                 </Pressable>
             </View>
@@ -47,15 +57,22 @@ const styles = StyleSheet.create({
     ScreenContainer: {
         display: "flex",
         height: "100%",
-        justifyContent: 'center',
         flexDirection: 'column'
     },
     ContentContainer: {
         paddingHorizontal: width / 16,
         display: 'flex',
         flexDirection: 'column',
+        flexGrow: 1,
         alignItems: 'center',
+        justifyContent: 'center',
     },
+    BackButton: {
+        width: 24,
+        height: 24,
+        marginLeft: 20,
+        marginBottom: 10
+    }
 })